+++ crates/zng-color/src/lib.rs
@@ -10,13 +10,15 @@
 #![warn(missing_docs)]
 #![recursion_limit = "256"]
 
-use std::{fmt, sync::Arc};
+use std::{fmt, ops, sync::Arc};
 use zng_app_context::context_local;
 
-use zng_layout::unit::{about_eq, about_eq_hash, AngleDegree, Factor};
+use zng_layout::unit::{about_eq, about_eq_hash, AngleDegree, Factor, FactorUnits};
 use zng_var::{
     animation::{easing::EasingStep, Transition, Transitionable},
-    context_var, impl_from_and_into_var, merge_var, IntoVar, Var, VarValue,
+    context_var, expr_var, impl_from_and_into_var,
+    types::ContextualizedVar,
+    IntoVar, Var, VarValue,
 };
 
 pub use zng_view_api::config::ColorScheme;
@@ -826,93 +828,248 @@ context_var! {
     pub static COLOR_SCHEME_VAR: ColorScheme = ColorScheme::default();
 }
 
-/// Create a variable that maps to `dark` or `light` depending on the contextual [`COLOR_SCHEME_VAR`].
-pub fn color_scheme_map<T: VarValue>(dark: impl IntoVar<T>, light: impl IntoVar<T>) -> impl Var<T> {
-    merge_var!(COLOR_SCHEME_VAR, dark.into_var(), light.into_var(), |&scheme, dark, light| {
-        match scheme {
-            ColorScheme::Dark => dark.clone(),
-            ColorScheme::Light => light.clone(),
-        }
-    })
-}
-
-/// Create a variable that selects the [`ColorPair`] depending on the contextual [`COLOR_SCHEME_VAR`].
-pub fn color_scheme_pair(pair: impl IntoVar<ColorPair>) -> impl Var<Rgba> {
-    merge_var!(COLOR_SCHEME_VAR, pair.into_var(), |&scheme, &pair| {
-        match scheme {
-            ColorScheme::Dark => pair.dark,
-            ColorScheme::Light => pair.light,
-        }
-    })
-}
-
-/// Create a variable that selects the [`ColorPair`] highlight depending on the contextual [`COLOR_SCHEME_VAR`].
-pub fn color_scheme_highlight(pair: impl IntoVar<ColorPair>, highlight: impl IntoVar<Factor>) -> impl Var<Rgba> {
-    merge_var!(
-        COLOR_SCHEME_VAR,
-        pair.into_var(),
-        highlight.into_var(),
-        |&scheme, &pair, &highlight| {
-            match scheme {
-                ColorScheme::Dark => pair.highlight_dark(highlight),
-                ColorScheme::Light => pair.highlight_light(highlight),
-            }
-        }
-    )
+/// RGBA color pair.
+///
+/// # Arguments
+///
+/// The arguments can be any color type that converts to [`Rgba`]. The first color
+/// is used in [`ColorScheme::Dark`] contexts, the second color is used in [`ColorScheme::Light`] contexts.
+///
+/// Note that [`LightDark`] converts `IntoVar<Rgba>` with a contextual var that selects the color, so you
+/// can just set color properties directly with a color pair .
+pub fn light_dark(light: impl Into<Rgba>, dark: impl Into<Rgba>) -> LightDark {
+    LightDark {
+        light: light.into(),
+        dark: dark.into(),
+    }
 }
 
 /// Represents a dark and light *color*.
+///
+///
+/// Note that this struct converts `IntoVar<Rgba>` with a contextual var that selects the color, so you
+/// can just set color properties directly with a color pair.
 #[derive(Debug, Clone, Copy, PartialEq, Hash, serde::Serialize, serde::Deserialize, Transitionable)]
-pub struct ColorPair {
+pub struct LightDark {
     /// Color used when [`ColorScheme::Dark`].
     pub dark: Rgba,
     /// Color used when [`ColorScheme::Light`].
     pub light: Rgba,
 }
 impl_from_and_into_var! {
-    /// From `(dark, light)` tuple.
-    fn from<D: Into<Rgba>, L: Into<Rgba>>((dark, light): (D, L)) -> ColorPair {
-        ColorPair {
-            dark: dark.into(),
-            light: light.into(),
-        }
+    // /// From `(light, dark)` tuple. !!: TODO
+    // fn from<L: Into<Rgba>, D: Into<Rgba>>((light, dark): (L, D)) -> LightDark {
+    //     LightDark {
+    //         light: light.into(),
+    //         dark: dark.into(),
+    //     }
+    // }
+
+    /// From same color to both.
+    fn from(color: Rgba) -> LightDark {
+        LightDark { dark: color, light: color }
     }
 
     /// From same color to both.
-    fn from(color: Rgba) -> ColorPair {
-        ColorPair { dark: color, light: color }
-    }
-
-    /// From same color to both.
-    fn from(color: Hsva) -> ColorPair {
+    fn from(color: Hsva) -> LightDark {
         Rgba::from(color).into()
     }
 
     /// From same color to both.
-    fn from(color: Hsla) -> ColorPair {
+    fn from(color: Hsla) -> LightDark {
         Rgba::from(color).into()
     }
 }
-impl ColorPair {
-    /// Overlay white with `highlight` amount as alpha over the [`dark`] color.
-    ///
-    /// [`dark`]: ColorPair::dark
-    pub fn highlight_dark(self, highlight: impl Into<Factor>) -> Rgba {
-        colors::WHITE.with_alpha(highlight.into()).mix_normal(self.dark)
-    }
+impl IntoVar<Rgba> for LightDark {
+    type Var = ContextualizedVar<Rgba>;
 
-    /// Overlay black with `highlight` amount as alpha over the [`light`] color.
-    ///
-    /// [`light`]: ColorPair::light
-    pub fn highlight_light(self, highlight: impl Into<Factor>) -> Rgba {
-        colors::BLACK.with_alpha(highlight.into()).mix_normal(self.light)
-    }
-
-    /// Gets the color for the scheme.
-    pub fn color(self, scheme: ColorScheme) -> Rgba {
-        match scheme {
+    fn into_var(self) -> Self::Var {
+        COLOR_SCHEME_VAR.map(move |s| match s {
             ColorScheme::Light => self.light,
             ColorScheme::Dark => self.dark,
+        })
+    }
+}
+impl LightDark {
+    /// New from light, dark colors.
+    pub fn new(light: impl Into<Rgba>, dark: impl Into<Rgba>) -> Self {
+        Self {
+            light: light.into(),
+            dark: dark.into(),
+        }
+    }
+
+    /// Overlay WHITE/BLACK to the dark/light color depending on the `factor`, negative factor inverts overlay.
+    pub fn shade_fct(mut self, factor: impl Into<Factor>) -> Self {
+        let mut factor = factor.into();
+        let (dark_overlay, light_overlay) = if factor > 0.fct() {
+            (colors::WHITE, colors::BLACK)
+        } else {
+            factor = factor.abs();
+            (colors::BLACK, colors::WHITE)
+        };
+        self.dark = dark_overlay.with_alpha(factor).mix_normal(self.dark);
+        self.light = light_overlay.with_alpha(factor).mix_normal(self.light);
+        self
+    }
+
+    /// Shade at 8% increments.
+    ///
+    /// Common usage: 1=hovered, 2=pressed.
+    pub fn shade(self, shade: i8) -> Self {
+        self.shade_fct(shade as f32 * 0.08)
+    }
+
+    /// Gets a contextual `Rgba` var that selects the color for the context scheme.
+    ///
+    /// Also see [`LightDarkVarExt`] for mapping from vars.
+    pub fn rgba(self) -> ContextualizedVar<Rgba> {
+        IntoVar::<Rgba>::into_var(self)
+    }
+
+    /// Gets a contextual `Rgba` var that selects the color for the context scheme and `map` it.
+    pub fn rgba_map<T: VarValue>(self, mut map: impl FnMut(Rgba) -> T + Send + 'static) -> impl Var<T> {
+        COLOR_SCHEME_VAR.map(move |s| match s {
+            ColorScheme::Light => map(self.light),
+            ColorScheme::Dark => map(self.dark),
+        })
+    }
+
+    /// Gets a contextual `Rgba` var that selects the color for the context scheme and converts it to `T`.
+    pub fn rgba_into<T: VarValue + From<Rgba>>(self) -> impl Var<T> {
+        self.rgba_map(T::from)
+    }
+}
+impl ops::Index<ColorScheme> for LightDark {
+    type Output = Rgba;
+
+    fn index(&self, index: ColorScheme) -> &Self::Output {
+        match index {
+            ColorScheme::Light => &self.light,
+            ColorScheme::Dark => &self.dark,
+        }
+    }
+}
+impl ops::IndexMut<ColorScheme> for LightDark {
+    fn index_mut(&mut self, index: ColorScheme) -> &mut Self::Output {
+        match index {
+            ColorScheme::Light => &mut self.light,
+            ColorScheme::Dark => &mut self.dark,
+        }
+    }
+}
+
+/// Extension methods for `impl Var<LightDark>`.
+pub trait LightDarkVarExt {
+    /// Gets a contextualized var that maps to [`LightDark::rgba`].
+    fn rgba(&self) -> impl Var<Rgba>;
+    /// Gets a contextualized var that maps to [`LightDark::rgba`] and `map`.
+    fn rgba_map<T: VarValue>(&self, map: impl FnMut(Rgba) -> T + Send + 'static) -> impl Var<T>;
+    /// Gets a contextualized var that maps to [`LightDark::rgba`] converted into `T`.
+    fn rgba_into<T: VarValue + From<Rgba>>(&self) -> impl Var<T>;
+
+    /// Gets a contextualized var that maps using `map` and then to [`LightDark::rgba`].
+    fn map_rgba(&self, map: impl FnMut(LightDark) -> LightDark + Send + 'static) -> impl Var<Rgba>;
+    /// Gets a contextualized var that maps using `map` and then into `T`.
+    fn map_rgba_into<T: VarValue + From<Rgba>>(&self, map: impl FnMut(LightDark) -> LightDark + Send + 'static) -> impl Var<T>;
+
+    /// Gets a contextualized var that maps to [`LightDark::shade_fct`] and then to [`LightDark::rgba`].
+    fn shade_fct(&self, fct: impl Into<Factor>) -> impl Var<Rgba>;
+    /// Gets a contextualized var that maps to [`LightDark::shade_fct`] and then to [`LightDark::rgba`] and then into `T`.
+    fn shade_fct_into<T: VarValue + From<Rgba>>(&self, fct: impl Into<Factor>) -> impl Var<T>;
+
+    /// Gets a contextualized var that maps to [`LightDark::shade`] and then to [`LightDark::rgba`].   
+    ///
+    /// * +1 - Hovered.
+    /// * +2 - Pressed.
+    fn shade(&self, shade: i8) -> impl Var<Rgba>;
+    /// Gets a contextualized var that maps to [`LightDark::shade`] and then to [`LightDark::rgba`] and then into `T`.
+    fn shade_into<T: VarValue + From<Rgba>>(&self, shade: i8) -> impl Var<T>;
+}
+impl<V: Var<LightDark>> LightDarkVarExt for V {
+    fn rgba(&self) -> impl Var<Rgba> {
+        expr_var! {
+            let c = #{self.clone()};
+            match *#{COLOR_SCHEME_VAR} {
+                ColorScheme::Light => c.light,
+                ColorScheme::Dark => c.dark,
+            }
+        }
+    }
+
+    fn rgba_map<T: VarValue>(&self, mut map: impl FnMut(Rgba) -> T + Send + 'static) -> impl Var<T> {
+        expr_var! {
+            let c = #{self.clone()};
+            match *#{COLOR_SCHEME_VAR} {
+                ColorScheme::Light => map(c.light),
+                ColorScheme::Dark => map(c.dark),
+            }
+        }
+    }
+
+    fn rgba_into<T: VarValue + From<Rgba>>(&self) -> impl Var<T> {
+        self.rgba_map(Into::into)
+    }
+
+    fn map_rgba(&self, mut map: impl FnMut(LightDark) -> LightDark + Send + 'static) -> impl Var<Rgba> {
+        expr_var! {
+            let c = map(*#{self.clone()});
+            match *#{COLOR_SCHEME_VAR} {
+                ColorScheme::Light => c.light,
+                ColorScheme::Dark => c.dark,
+            }
+        }
+    }
+
+    fn map_rgba_into<T: VarValue + From<Rgba>>(&self, mut map: impl FnMut(LightDark) -> LightDark + Send + 'static) -> impl Var<T> {
+        expr_var! {
+            let c = map(*#{self.clone()});
+            match *#{COLOR_SCHEME_VAR} {
+                ColorScheme::Light => T::from(c.light),
+                ColorScheme::Dark => T::from(c.dark),
+            }
+        }
+    }
+
+    fn shade_fct(&self, fct: impl Into<Factor>) -> impl Var<Rgba> {
+        let fct = fct.into();
+        expr_var! {
+            let c = #{self.clone()}.shade_fct(fct);
+            match *#{COLOR_SCHEME_VAR} {
+                ColorScheme::Light => c.light,
+                ColorScheme::Dark => c.dark,
+            }
+        }
+    }
+
+    fn shade(&self, shade: i8) -> impl Var<Rgba> {
+        expr_var! {
+            let c = #{self.clone()}.shade(shade);
+            match *#{COLOR_SCHEME_VAR} {
+                ColorScheme::Light => c.light,
+                ColorScheme::Dark => c.dark,
+            }
+        }
+    }
+
+    fn shade_fct_into<T: VarValue + From<Rgba>>(&self, fct: impl Into<Factor>) -> impl Var<T> {
+        let fct = fct.into();
+        expr_var! {
+            let c = #{self.clone()}.shade_fct(fct);
+            match *#{COLOR_SCHEME_VAR} {
+                ColorScheme::Light => T::from(c.light),
+                ColorScheme::Dark => T::from(c.dark),
+            }
+        }
+    }
+
+    fn shade_into<T: VarValue + From<Rgba>>(&self, shade: i8) -> impl Var<T> {
+        expr_var! {
+            let c = #{self.clone()}.shade(shade);
+            match *#{COLOR_SCHEME_VAR} {
+                ColorScheme::Light => T::from(c.light),
+                ColorScheme::Dark => T::from(c.dark),
+            }
         }
     }
 }
@@ -981,7 +1138,7 @@ context_local! {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use zng_layout::unit::{AngleUnits as _, FactorUnits as _};
+    use zng_layout::unit::AngleUnits as _;
 
     #[test]
     fn hsl_red() {