+++ crates/zng/src/button.rs
@@ -88,7 +88,7 @@
 //!
 //! ## Base Colors
 //!
-//! The default style derive all colors from the [`base_colors`](fn@base_colors), so if you
+//! The default style derive all colors from the [`base_color`](fn@crate::color::base_color), so if you
 //! only want to change color of buttons you can use this property.
 //!
 //! The example below extends the button style to change the button color to red when it represents
@@ -96,17 +96,17 @@
 //!
 //! ```
 //! use zng::prelude::*;
-//! use zng::button;
+//! use zng::{button, color::base_color};
 //!
 //! # let _scope = APP.defaults(); let _ =
 //! Window! {
 //!     button::style_fn = Style! {
 //!         when *#{button::BUTTON.cmd()} == Some(window::cmd::CLOSE_CMD) {
-//!             button::base_colors = color::ColorPair {
-//!                 // dark theme base
-//!                 dark: colors::BLACK.with_alpha(80.pct()).mix_normal(colors::RED),
+//!             base_color = color::LightDark {
 //!                 // light theme base
 //!                 light: colors::WHITE.with_alpha(80.pct()).mix_normal(colors::RED),
+//!                 // dark theme base
+//!                 dark: colors::BLACK.with_alpha(80.pct()).mix_normal(colors::RED),
 //!             };
 //!         }
 //!     };
@@ -118,4 +118,4 @@
 //!
 //! See [`zng_wgt_button`] for the full widget API.
 
-pub use zng_wgt_button::{base_colors, style_fn, Button, DefaultStyle, LightStyle, LinkStyle, BUTTON};
+pub use zng_wgt_button::{style_fn, Button, DefaultStyle, LightStyle, LinkStyle, BUTTON};