+++ examples/scroll/src/main.rs
@@ -20,9 +20,9 @@ fn main() {
             child = Scroll! {
                 id = "scroll";
                 padding = 20;
-                widget::background_color = color::color_scheme_map(
+                widget::background_color = light_dark(
+                    colors::WHITE.with_alpha(80.pct()).mix_normal(hex!(#245E81)),
                     hex!(#245E81),
-                    colors::WHITE.with_alpha(80.pct()).mix_normal(hex!(#245E81))
                 );
                 smooth_scrolling = smooth_scrolling.map_into();